Fix pg query count datatype
node-postgres returns numeric fields as string unless it is clear that
these can be converted lossless to JS Number. Cast the count to ::float
(double precission by default) to match the JS type.

Resolves: #5528

+++ packages/server/graphql/intranetSchema/queries/dailyPulse.ts
@@ -38,7 +38,7 @@ const filterCounts = (domainCount: DomainCount[]) =>
 const addAllTimeTotals = async (domainCount: DomainCount[]): Promise<DomainCountWithAllTime[]> => {
   const pg = getPg()
   const allTimeCount = await pg.query(
-    `SELECT count(*) as "allTimeTotal", split_part(email, '@', 2) as domain from "User"
+    `SELECT count(*)::float as "allTimeTotal", split_part(email, '@', 2) as domain from "User"
      WHERE split_part(email, '@', 2) = ANY($1::text[])
      GROUP BY split_part(email, '@', 2)`,
     [domainCount.map((count) => count.domain)]
